lambda/requirements.txt
crhelper
boto3
httpretty
kafka-python
